Rename proteus-events to proteus-orchestrate
