--- proteus-events/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/thetorproject/proteus/proteus-events/cmd"
-
-func main() {
-	cmd.Execute()
-}
