+++ proteus-orchestrate/events/bindata.go
