+++ proteus-orchestrate/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/thetorproject/proteus/proteus-orchestrate/cmd"
+
+func main() {
+	cmd.Execute()
+}
